Working model, periods and fruits.
Model now works. We can pass in new demand columns, calling them new periods. We can build the model up and test for robustness. I've checked the flows for one of the avocado flows and it all checks out. I am assuming it works for the rest of the flows and fruit types. Keep the set and parameter names. We can append to the demand colomn and it should all feed through. I'll explain in the lab tomorrow - 14/08/2018.